nitial TypeScript support for the project. Major changes: -All source files now live in /src -All transpiled javascript files will appear in /build -Moved all non-source files (images) into new /resources directory
All source files will be renamed to .ts or .tsx and type markups will be added. For now I only converted a few files as example. Converting more files is a lot more work as there are errors to resolve.

All declaration files are installed using the nom @types system. Sea re not using typings at all.

New files:
tsconfig.json is configuration for TypeScript.

New dependencies:
-Pay attention to package.json for dependency details.
-New dev dependency to several TypeScript declaration modules @types: @types/react, @types/react-native, @types/graphql and @types/redux-promise.
-@graphql also pulls in dependency to async and isomorphic-tech.
New dev dependency to Typescript compiler.

Longer dependency discussion:
Since we are switching to essentially a different language (TypeSCript) this is a hard dependency on the transpolar (tic) and the new language format.
If in the future we wanted to go to pure javascript, we need to remove all the typings from the code, uninstall the @types nom’s, uninstall the typescript compiler, the the project would function as normal.

There is a different way to implement this. We can implement all our code in raw javascript and provide a index.d.ts declaration of our library. This would allow us to keep the typings separate from the code. This would make code easier to read for javascript programmers, but slower and harder to develop since we’d have to look in 2 places to understand types.

In the future this is certainly a possible change to make, it would be a refactor away. A major refactor - every line of code nearly would have to be changed.

But it is grunt work, and it can be done piecewise.

It is a subtle but important difference. Do we see ourselves as a library that tries to be light weight, or as a full service framework that will always be the primary building block?